move from_network to bokeh.plotting
