+++ sphinx/source/docs/user_guide/examples/graph_interaction.py
@@ -3,8 +3,8 @@ import networkx as nx
 from bokeh.io import output_file, show
 from bokeh.models import (BoxSelectTool, Circle, EdgesAndLinkedNodes, HoverTool,
                           MultiLine, NodesAndLinkedEdges, Plot, Range1d, TapTool)
-from bokeh.models.graphs import from_networkx
 from bokeh.palettes import Spectral4
+from bokeh.plotting import from_networkx
 
 G=nx.karate_club_graph()
