+++ sphinx/source/docs/user_guide/examples/graph_networkx.py
@@ -1,8 +1,7 @@
 import networkx as nx
 
 from bokeh.io import show, output_file
-from bokeh.models.graphs import from_networkx
-from bokeh.plotting import figure
+from bokeh.plotting import figure, from_networkx
 
 G = nx.karate_club_graph()
